Pants-ify the betarepo.
Just for the hell of it.

Anyway - this is just a little practice repo
for me. No plans for it - but I have lots of
little libraries and scripts that implement
trees or sorts for when I need them - I always
thought it would be nice to get them into genuine
library code so that they could be reused as
"known good".

It isn't fun to implement something complicated
only to find out you had a trivial logic bug in
the Node construction or something.

If I even semi-regularly dump copies of structures I
use elsewhere this would be a win for me.